Solution one
I have realized, I could make a real time version of this, not just one which takes in data and makes calculations, then displays the end result. A real time solution could read in one value at a time and make calculations on the current value which would be subject to change.